requirement.txt
apiclient==1.0.4
asgiref==3.5.2
boto3==1.26.102
botocore==1.29.102
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1
django-cors-headers==3.13.0
django-ipware==4.0.2
django-ratelimit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
mysqlclient==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==9.2.0
printwhatitype==0.0.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyttsx==1.1
pytz==2022.6
ratelimit==2.2.1
reportlab==3.6.11
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.15
